requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# PDF processing
PyMuPDF==1.23.8
fitz==0.0.1.dev2

# Embeddings and vector search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# LLM integration
requests==2.31.0

# Utilities
python-multipart==0.0.6
tqdm==4.66.1

# Development
python-dotenv==1.0.0